Made utils file for standard library
- Moved some functions that didn't really fit in `compat.py` there
