+++ nengo/utils/simulator.py
@@ -1,8 +1,9 @@
 from collections import defaultdict
 import itertools
 
-from nengo.utils.compat import iteritems, groupby
+from nengo.utils.compat import iteritems
 from nengo.utils.graphs import add_edges
+from nengo.utils.stdlib import groupby
 
 
 def operator_depencency_graph(operators):  # noqa: C901
