+++ nengo/utils/tests/test_stdlib.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 import nengo
-from nengo.utils.inspect import checked_call
+from nengo.utils.stdlib import checked_call, groupby
 
 logger = logging.getLogger(__name__)
 
@@ -68,6 +68,41 @@ def test_checked_call_errors():
     assert checked_call(np.sin, 1, 2, 3) == (None, False)
     with pytest.raises(ValueError):
         checked_call(lambda x: np.sin(1, 2, 3), 1)
+
+
+@pytest.mark.parametrize(
+    "hashable, force_list",
+    [(False, False), (False, True), (True, False), (True, True)])
+def test_groupby(hashable, force_list):
+    rng = np.random.RandomState(96)
+
+    if hashable:
+        keys = list(range(1, 5))
+    else:
+        keys = [[0, 0], [0, 1], [1, 0], [1, 1]]
+
+    keys = sorted(keys)
+
+    # make groups and pairs
+    groups = [rng.randn(rng.randint(5, 10)) for _ in keys]
+
+    pairs = []
+    for key, group in zip(keys, groups):
+        pairs.extend((key, value) for value in group)
+
+    # shuffle pairs
+    pairs = [pairs[i] for i in rng.permutation(len(pairs))]
+
+    # call groupby
+    keygroups = groupby(pairs, lambda p: p[0], force_list=force_list)
+
+    keys2 = sorted(map(lambda x: x[0], keygroups))
+    assert keys2 == keys
+
+    for key2, keygroup2 in keygroups:
+        group = groups[keys.index(key2)]
+        group2 = map(lambda x: x[1], keygroup2)
+        assert sorted(group2) == sorted(group)
 
 
 if __name__ == "__main__":
